+++ frontend/app/utils/sumMapBy.js
@@ -1,6 +1,6 @@
 export default function sumMapBy(array) {
   return array.reduce((acc, value) => {
     if (value === undefined) return acc;
-    return acc + parseInt(value);
+    return acc + parseFloat(value);
   }, 0);
 }
